+++ test/unit/mysql_rep_cmp/create_data_config.py
@@ -17,6 +17,7 @@
 import sys
 import os
 import unittest
+import mock
 
 # Local
 sys.path.append(os.getcwd())
@@ -50,7 +51,7 @@ class ArgParser(object):
 
         self.args_array = {
             "-c": "mysql_cfg", "-d": "config", "-e": "to_addr",
-            "-o": "outfile", "-n": "indentation", "-m": "mongo",
+            "-o": "outfile", "-n": "indentation",
             "-i": "database:table", "-w": "a", "-p": False}
 
     def get_val(self, skey, def_val=None):
@@ -74,6 +75,8 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
+        test_with_no_mongo
+        test_with_mongo
         test_mode_with_data
         test_mailx_with_no_data
         test_subj_with_no_data
@@ -92,12 +95,46 @@ class UnitTest(unittest.TestCase):
         """
 
         self.args = ArgParser()
+        self.args.args_array["-m"] = "mongo"
+        self.args2 = ArgParser()
         self.results = None
         self.results2 = "to_addr"
         self.results3 = False
         self.results4 = "a"
 
-    def test_mode_with_data(self):
+    def test_with_no_mongo(self):
+
+        """Function:  test_with_no_mongo
+
+        Description:  Test with mongo being passed.
+
+        Arguments:
+
+        """
+
+        data_config = mysql_rep_cmp.create_data_config(self.args2)
+
+        self.assertTrue("mongo" not in data_config)
+
+    @mock.patch("mysql_rep_cmp.gen_libs.load_module")
+    def test_with_mongo(self, mock_load):
+
+        """Function:  test_with_mongo
+
+        Description:  Test with mongo being passed.
+
+        Arguments:
+
+        """
+
+        mock_load.return_value = "Mongo_Config"
+
+        self.assertEqual(
+            mysql_rep_cmp.create_data_config(self.args)["mode"],
+            self.results4)
+
+    @mock.patch("mysql_rep_cmp.gen_libs.load_module")
+    def test_mode_with_data(self, mock_load):
 
         """Function:  test_mode_with_data
 
@@ -107,11 +144,14 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        mock_load.return_value = "Mongo_Config"
+
         self.assertEqual(
             mysql_rep_cmp.create_data_config(self.args)["mode"],
             self.results4)
 
-    def test_mailx_with_no_data(self):
+    @mock.patch("mysql_rep_cmp.gen_libs.load_module")
+    def test_mailx_with_no_data(self, mock_load):
 
         """Function:  test_mailx_with_no_data
 
@@ -121,11 +161,14 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        mock_load.return_value = "Mongo_Config"
+
         self.assertEqual(
             mysql_rep_cmp.create_data_config(self.args)["mailx"],
             self.results3)
 
-    def test_subj_with_no_data(self):
+    @mock.patch("mysql_rep_cmp.gen_libs.load_module")
+    def test_subj_with_no_data(self, mock_load):
 
         """Function:  test_subj_with_no_data
 
@@ -135,11 +178,14 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        mock_load.return_value = "Mongo_Config"
+
         self.assertEqual(
             mysql_rep_cmp.create_data_config(self.args)["subj"],
             self.results)
 
-    def test_to_addr_with_data(self):
+    @mock.patch("mysql_rep_cmp.gen_libs.load_module")
+    def test_to_addr_with_data(self, mock_load):
 
         """Function:  test_to_addr_with_data
 
@@ -148,6 +194,8 @@ class UnitTest(unittest.TestCase):
         Arguments:
 
         """
+
+        mock_load.return_value = "Mongo_Config"
 
         self.assertEqual(
             mysql_rep_cmp.create_data_config(self.args)["to_addr"],